[#29] refactor(view) - template 분리 구조 적용
- 기존 뷰의 이름과 동일한 폴더를 생성
- 폴더 밑에 index.ts에 기존 뷰의 내용을 넣음
- template.js에 뷰가 가지고 있던 template literal 코드를 넣음
- 스타일 작업을 위한 style.scss 추가 및 연결
