+++ frontend/view/ContainerView/RouterView/CalendarView/template.ts
@@ -1,0 +1,6 @@
+export const template: string = `
+  <section id='invoice-calendar'>
+    <div class="rows">
+    </div>
+  </section>
+`
