+++ frontend/view/ContainerView/RouterView/ChartView/style.scss
@@ -1,0 +1,1 @@
+@use '@/static/styles/package' as *;
