+++ frontend/view/ContainerView/RouterView/ChartView/template.ts
@@ -1,0 +1,7 @@
+export const template: string = `
+  <section id='invoice-chart'>
+    <div>
+      차트입니당!
+    </div>
+  </section>
+`
