+++ frontend/view/ContainerView/RouterView/FilterView/template.ts
@@ -1,0 +1,14 @@
+export const template: string = `
+  <section id='invoice-filter'>
+    <div>
+      <input type="checkbox" class="earning-checkbox" />
+      <label>수입</label>
+      <label class="earning-total">3,234,123원</label>
+    </div>
+    <div>
+      <input type="checkbox" class="spending-checkbox" />
+      <label>지출</label>
+      <label class="spending-total">134,123원</label>
+    </div>
+  </section>
+`
