+++ frontend/view/ContainerView/RouterView/template.ts
@@ -1,0 +1,4 @@
+export const template: string = `
+  <div id='router-content'>
+  </div>
+`
