+++ frontend/view/MainView/HeaderView/index.ts
@@ -1,13 +1,6 @@
-import { View } from '../index'
-
-const template = `
-  <header id="header">
-    <div class="row">
-      <div class="item center main-title">룰루랄라 가계부</div>
-      <button class="item right button-payment-modal">결제 수단 관리</button>
-    </div>
-  </header>
-`
+import { View } from '../../index'
+import './style.scss'
+import { template } from './template'
 
 export default class HeaderView extends View {
   $buttonPaymentModal: HTMLButtonElement
