+++ frontend/view/MainView/HeaderView/style.scss
@@ -1,0 +1,1 @@
+@use '@/static/styles/package' as *;
