+++ frontend/view/MainView/HeaderView/template.ts
@@ -1,0 +1,8 @@
+export const template: string = `
+  <header id="header">
+    <div class="row">
+      <div class="item center main-title">룰루랄라 가계부</div>
+      <button class="item right button-payment-modal">결제 수단 관리</button>
+    </div>
+  </header>
+`
