+++ frontend/view/MainView/NavigatorView/index.ts
@@ -1,0 +1,10 @@
+import { View } from '../../index'
+import './style.scss'
+import { template } from './template'
+
+export default class NavigatorView extends View {
+  constructor() {
+    super(template)
+  }
+  mount(): void {}
+}
