+++ frontend/view/MainView/template.ts
@@ -1,0 +1,4 @@
+export const template: string = `
+  <main id='main'>
+  </main>
+`
